+++ CHANGELOG.md
@@ -78,6 +78,10 @@ Full documentation for ROCm Compute Profiler is available at [https://rocm.docs.
 * Analysis of new workloads might require providing shader/memory clock speed using
 --specs-correction operation if `amd-smi` or `rocminfo` does not provide clock speeds.
 
+### Removed
+
+* Roofline support for Ubuntu 20.04 and SLES below 15.6
+
 ## ROCm Compute Profiler 3.1.0 for ROCm 6.4.0
 
 ### Added